% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-homerange.R, R/homerange-functions.R
\name{area.BRB}
\alias{area.BRB}
\alias{area.BRB}
\title{Updated BRB code}
\usage{
area.BRB(x = NULL, start.date = NULL, end.date = NULL, hab = NULL,
  ndvi = NULL, iso = 95, plot.it = FALSE, t = "UD", vv = NULL)

area.BRB(x = NULL, start.date = NULL, end.date = NULL, hab = NULL,
  ndvi = NULL, iso = 95, plot.it = FALSE, t = "UD", vv = NULL)
}
\arguments{
\item{x}{definition}

\item{start.date}{definition}

\item{end.date}{definition}

\item{hab}{definition}

\item{iso}{definition}

\item{plot.it}{definition}

\item{t}{definition}

\item{vv}{definition}

\item{x}{definition}

\item{start.date}{definition}

\item{end.date}{definition}

\item{hab}{definition}

\item{iso}{definition}

\item{plot.it}{definition}

\item{t}{definition}

\item{vv}{definition}
}
\description{
Updated BRB code

For the original function, see adehabitatHR::BRB
}
\examples{


}
