% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{biomass_avail_hr}
\alias{biomass_avail_hr}
\title{Calcuate available biomass per homerange using the indices as weights.}
\usage{
biomass_avail_hr(group, period_start_date, period_end_date,
  group_period_label, hr_type_incl = 95, hr_sf, tr_geom, tr_pheno_fpv,
  indices, figs_sf)
}
\arguments{
\item{group}{def}

\item{period_start_date}{def}

\item{period_end_date}{def}

\item{group_period_label}{def}

\item{hr_type_incl}{50, 70, or 95}

\item{hr_sf}{def}

\item{tr_geom}{Transect data with geometries}

\item{tr_pheno_fpv}{Use table including geometries}

\item{indices}{Index data to use as weights created by pheno_avail_indices_sr}

\item{figs_sf}{def}
}
\description{
Calcuate available biomass per homerange using the indices as weights.
}
\examples{
biomass_hr_first <- biomass_hr(group = "CP", period_start_date = as.Date("2009-08-24"),
                               period_end_date = as.Date("2009-12-11"), group_period_label = "CP_1",
                               hr_type_incl = 95, hr_sf, tr_geom, tr_pheno_fpv, indices, figs_sf)
}
