% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{biomass_avail_sr}
\alias{biomass_avail_sr}
\title{Calcuate available biomass using the indices as weights.}
\usage{
biomass_avail_sr(biomass_max = NULL, indices = NULL)
}
\arguments{
\item{biomass_max}{Dataframe of max biomass data created by biomass_max_sr}

\item{indices}{Index data to use as weights created by pheno_avail_indices_sr}
}
\description{
Calcuate available biomass using the indices as weights.
}
\examples{
biomass_avail_lo <- biomass_avail_sr(biomass_max, indices_lo)
}

