% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{biomass_max_sr}
\alias{biomass_max_sr}
\title{Calculate max potential biomass for each species (if availability were 1).}
\usage{
biomass_max_sr(df = NULL)
}
\arguments{
\item{df}{Dataframe of data for biomass calculation created by transect_subset_sr}
}
\description{
Calculate max potential biomass for each species (if availability were 1).
}
\examples{
biomass_max <- biomass_max_sr(tr_pheno_fpv)
}

