% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{biomass_monthly_summary}
\alias{biomass_monthly_summary}
\title{Summarize available biomass data by summing species in each month / year.}
\usage{
biomass_monthly_summary(df = NULL)
}
\arguments{
\item{df}{Dataframe of available biomass data created by biomass_avail_sr}
}
\description{
Summarize available biomass data by summing species in each month / year.
}
\examples{
biomass_summary_lo <- biomass_monthly_summary(biomass_avail_lo)
}
