% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks_connect.R
\name{databricks_connect}
\alias{databricks_connect}
\title{databricks_connect}
\usage{
databricks_connect()
}
\value{
connection to Habitat's Databricks database
}
\description{
This function sets up a connection to Habitat's Databricks database. It requires that you have correctly
saved environment variables needed to connect to Databricks and installed the correct JDBC driver and
Java software on your computer. For a guide to these installations see the the team's \href{https://habitathealth.atlassian.net/wiki/spaces/teamfaa190968e2a4b35963a58c1a4cb3842/pages/161579009/Analytics+in+R}{Analytics in R}
page
}
\examples{
\dontrun{
run_sql_string(con = databricks_connect(),
               sql_string = 'SELECT * FROM staging.de_core.stg_census_sac')}
}
