% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{figs_to_sf}
\alias{figs_to_sf}
\title{Transforms fig table into a sf-dataframe}
\usage{
figs_to_sf(figs)
}
\arguments{
\item{figs}{Object with fig data (e.g. table from FicusData_Nov13_2013.csv)}
}
\description{
Transforms fig table into a sf-dataframe
}
