% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{fpv_get_min_dbh_sr}
\alias{fpv_get_min_dbh_sr}
\title{Set minimum DBHs for fruit-producing trees of each species.}
\usage{
fpv_get_min_dbh_sr(fpv = NULL)
}
\arguments{
\item{fpv}{The FPV data.}
}
\description{
Set minimum DBHs for fruit-producing trees of each species.
}
\examples{
min_dbh <- fpv_get_min_dbh_sr(fpv)
}
