% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{fpv_subset_pheno_sr}
\alias{fpv_subset_pheno_sr}
\title{Get relevant FPV data corresponding to pheno species}
\usage{
fpv_subset_pheno_sr(fpv = NULL, pheno = NULL)
}
\arguments{
\item{fpv}{FPV data.}

\item{pheno}{Phenology data created by pheno_prep_fruit_sr.}
}
\description{
Get relevant FPV data corresponding to pheno species
}
\examples{
fpv <- fpv_subset_pheno_sr(fpv, pheno)
}

