% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load-pace-tables.R
\name{get_alphamale_tenures}
\alias{get_alphamale_tenures}
\title{Get the Alphamale-tenure table.}
\usage{
get_alphamale_tenures(pace_db, full = TRUE)
}
\arguments{
\item{pace_db}{The src_mysql connection to the PACE Database.}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}
}
\description{
Get the Alphamale-tenure table.
}
\examples{
get_alphamale_tenures(pace_db)
}

