% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-pace-life-history-tables.R
\name{get_biography}
\alias{get_biography}
\title{Get table with biography from birth until death, disappearance or end of observation of individuals.}
\usage{
get_biography(paceR_db, full = TRUE, projectID = 1)
}
\arguments{
\item{paceR_db}{The src_mysql connection to the paceR Database (view-collection).}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}

\item{projectID}{Option to get data only from specific project (1-7?). Default is 1 (Santa Rosa)}
}
\description{
Get table with biography from birth until death, disappearance or end of observation of individuals.
}
\examples{
get_biography(paceR_db)
}

