% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{get_biomass_sr}
\alias{get_biomass_sr}
\title{Calculate available biomass. This function is a wrapper for most of the other phenology functions.}
\usage{
get_biomass_sr(ph = NULL, tr = NULL, fpv = NULL, figs = NULL,
  exclude_species = "", smooth = "none")
}
\arguments{
\item{ph}{Dataframe of raw PACE phenology data.}

\item{tr}{Dataframe of raw PACE transect data.}

\item{fpv}{Dataframe of FPV data (from CSV file).}

\item{exclude_species}{A character vector of species codes to exclude.}

\item{smooth}{Either "none", "loess", or "gam. The default is "none".}
}
\description{
Calculate available biomass. This function is a wrapper for most of the other phenology functions.
}
\examples{
exclude_species <- c("SCAP", "SPAV", "CCAN", "BUNG", "HCOU", "ATIB", "GULM", "LCAN", "LSPE", "FUNK")
biomass_avail_lo <- get_biomass_sr(ph, tr, fpv, exclude_species, smooth = "loess")
}
