% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standards.R
\name{get_census}
\alias{get_census}
\title{get_census}
\usage{
get_census(location = "all")
}
\arguments{
\item{location}{a string, one of 'all' (default), 'sac' for sacramento, or 'sla' for south LA}
}
\value{
tibble of census data depending on location choice
}
\description{
This function takes a Habitat Health location and pulls the entire census table
for that location. If \code{location = 'all'} is specified, all locations will be pulled
and appended to one another.
}
\examples{
\dontrun{
# get the census table data from the Sacramento location
get_census('sac')}
}
