% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load-pace-focaldata-mackenzie.R
\name{get_focaldata_MB}
\alias{get_focaldata_MB}
\title{Get focal behaviour recorded by Mackenzie Bergstrom for her MSc & PhD projects (SubProjectID = 13)}
\usage{
get_focaldata_MB(pace_db)
}
\arguments{
\item{pace_db}{The src_mysql connection to the PACE Database.}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}
}
\description{
Get focal behaviour recorded by Mackenzie Bergstrom for her MSc & PhD projects (SubProjectID = 13)
}
\examples{
get_biography(pace_db)
}

