% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-pace-focaldata-carnegie.R
\name{get_focaldata_SC}
\alias{get_focaldata_SC}
\title{Get focal behaviour recorded by Sarah Carnegie's for her PhD project (SubProjectID = 12)}
\usage{
get_focaldata_SC(paceR_db, full = TRUE)
}
\arguments{
\item{paceR_db}{The src_mysql connection to the paceR Database (view-collection).}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}
}
\description{
Get focal behaviour recorded by Sarah Carnegie's for her PhD project (SubProjectID = 12)
}
\examples{
get_focaldata_SC(paceR_db)
}

