% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerange-functions.R
\name{get_habitat_use}
\alias{get_habitat_use}
\title{Calculate habitat utilization from ranging data and topography (?) map}
\usage{
get_habitat_use(ranging_waypoints, start_date = NULL, ints_per_year = 1,
  hr_periods = NULL, min_nb_reloc = NULL)
}
\arguments{
\item{start_date}{Begin of first time interval. By default (NULL), Jan 1st of the earliest timestamp in ranging dataset - NOT YET IMPLEMENTED}

\item{ints_per_year}{Default is 1 (i.e. yearly homeranges). Set to NULL if not needed. Currently implemented:
\itemize{
 \item 12 (1 month-intervals)
 \item 8 (1.5 months-intervals)
 \item 4 (3 months-intervals)
 \item 2 (6 months-intervals)
 \item 1 (1 year-intervals)
 }}

\item{hr_periods}{If custom periods for home ranges are required (e.g. 05-02-2012 - 20-07-2012). Default is NULL.\cr
Create object of type: data.frame(date_begin = start_vec, date_end = end_vec)\cr
With start_vec including all start_dates and end_vec all end_dates\cr
Note: 'ints_per_year' is ignored if both are specified}

\item{min_nb_reloc}{Minimum number of location points per interval. Intervals for which the data are too sparse are removed.
  Default (NULL) uses Fernando's criteria:
  \itemize{
   \item 96 pts (~8 full days) for monthly home range
   \item 192 pts (~16 full days) for quarterly home range
   \item 384 pts (~32 full days) for half-yearly home range
   \item 768 pts (~64 full days) for annual home range
}}

\item{ranging_waypoint}{Table with ranging waypoint - MORE SPECIFICS}
}
\description{
Creates dataframe that contains various information on habitat use for each group and time interval\cr
spdf_50, spdf_70, spdf_95 contain objects of class 'SpatialPolygonDataFrame' (see package 'sp' for futher information)\cr
ud contains object of class 'estUD' (see 'adehabitatHR' for further information)\cr
Note: use intervals that start on first of each month when using with fruit biomass calculation (because of monthly phenology data)
}
\examples{

}
