% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-homerange.R
\name{get_homerange}
\alias{get_homerange}
\title{Calculate home ranges for time intervals}
\usage{
get_homerange(ranging_waypoints, start_date = NULL, ints_per_year = 1,
  hr_periods = NULL, min_nb_reloc = NULL, output = c("ud", "hr", "ob"))
}
\arguments{
\item{start_date}{Begin of first time interval. By default (NULL), Jan 1st of the earliest timestamp in ranging dataset - NOT YET IMPLEMENTED}

\item{ints_per_year}{Default is 1 (i.e. yearly homeranges). Set to NULL if not needed. Currently implemented:
12 (1 month-intervals)
8 (1.5 months-intervals)
4 (3 months-intervals)
2 (6 months-intervals)
1 (1 year-intervals)}

\item{hr_periods}{If custom periods for home ranges are required (e.g. 05-02-2012 - 20-07-2012). Default is NULL.
Create object of type: data.frame(date_begin = start_vec, date_end = end_vec)
With start_vec including all start_dates and end_vec all end_dates
  - Overwrites ints_per_year if both are specified}

\item{min_nb_reloc}{Minimum number of location points per interval. Intervals for which the data are too sparse are removed.
Default is NULL, which uses Fernando's criteria}

\item{output}{What type of object should be returned? ud, ob, or hr (sf-dataframe with homerange polygons)}

\item{ranging_waypoint}{Table with ranging waypoint - MORE SPECIFICS}
}
\description{
(Use intervals that start on first of month when used with phenology)
}
\examples{

}
