% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{get_horizontal_transects}
\alias{get_horizontal_transects}
\title{Get gps points for horizontal transects}
\usage{
get_horizontal_transects(tr_full, tr_pt)
}
\arguments{
\item{tr_full}{Dataframe with tblVegetationTransect_detailed from Pacelab/paceR}

\item{tr_pt}{Dataframe with 'tblVegetationTransectGridPoint' from Pacelab}

\item{paceR}{The src_mysql connection to the PACE Database.}
}
\description{
Get gps points for horizontal transects
}
