% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load-pace-tables.R
\name{get_individuals}
\alias{get_individuals}
\title{Get the Individuals table.}
\usage{
get_individuals(pace_db, full = TRUE)
}
\arguments{
\item{pace_db}{The src_mysql connection to the PACE Database.}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}
}
\description{
Get the Individuals table.
}
\examples{
get_individuals(pace_db)
}

