% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generate-pace-life-history-tables.R
\name{get_infanticide_risk}
\alias{get_infanticide_risk}
\title{Get table with risk of infanticide for infants
as a consequence of alpha male takeovers}
\usage{
get_infanticide_risk(pace_db, full = TRUE, projectID = 1)
}
\arguments{
\item{pace_db}{The src_mysql connection to the PACE Database.}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}

\item{projectID}{Option to get data only from specific project (1-7?), standard is set to 1 (Santa Rosa)}
}
\description{
Unstable (i.e. risky) periods start with the end of an alpha alpha male tenureship
and infants younger than 1 year are assumed to be at risk.
This does not count for infants conceived after the takeover
(i.e. born more than 5.5 months = 165 days after the takeover)
}
\examples{
get_infanticide_risk (pace_db)
}

