% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load-pace-tables.R
\name{get_monthly_census}
\alias{get_monthly_census}
\title{Get table with group-census data - New Version}
\usage{
get_monthly_census(paceR_db, full = TRUE)
}
\arguments{
\item{paceR_db}{The src_mysql connection to the paceR Database.}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}
}
\description{
Get table with group-census data - New Version
}
\examples{
get_monthly_census(paceR_db)
}

