% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-pace-focaldata-mackenzie.R
\name{get_proximity_MB}
\alias{get_proximity_MB}
\title{Get proximity scans recorded by Mackenzie Bergstrom for her MSc & PhD projects (SubProjectID = 13)}
\usage{
get_proximity_MB(paceR_db)
}
\arguments{
\item{paceR_db}{The src_mysql connection to the paceR Database (view-collection).}
}
\description{
Get proximity scans recorded by Mackenzie Bergstrom for her MSc & PhD projects (SubProjectID = 13)
}
\examples{
get_proximity_MB(paceR_db)
}
