% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerange-functions.R
\name{get_sf_from_hu}
\alias{get_sf_from_hu}
\title{Creates a 'simple feature' dataframe from the habitat utilization dataframe}
\usage{
get_sf_from_hu(habitat_use_df)
}
\arguments{
\item{habitat_use_df}{The dataframe returned from get_habitat_use()}
}
\description{
The returned object is a tidy (i.e. long) dataframe:\cr
One row per group, interval, and homerange-type (50, 70, and 95 percent)
}
\examples{

}
