% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{get_vertical_transects}
\alias{get_vertical_transects}
\title{Get gps points for vertical transects}
\usage{
get_vertical_transects(tr_full, data_dir = "data/")
}
\arguments{
\item{tr_full}{Dataframe with tblVegetationTransect_detailed from Pacelab/paceR}

\item{data_dir}{Define directory in which the file 'v_transfect_2016.gpx' can be found.}
}
\description{
Get gps points for vertical transects
}
