% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_paceR_views.R
\name{getv_AlphaMaleTenure}
\alias{getv_AlphaMaleTenure}
\title{Get the Alphamale-tenure table.}
\usage{
getv_AlphaMaleTenure(paceR_db, full = TRUE)
}
\arguments{
\item{paceR_db}{The src_mysql connection to the PaceR Database.}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}
}
\description{
Get the Alphamale-tenure table.
}
\examples{
getv_AlphaMaleTenure(paceR_db)
}

