% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_paceR_views.R
\name{getv_CensusMonthly}
\alias{getv_CensusMonthly}
\title{Get table with group-census data}
\usage{
getv_CensusMonthly(paceR_db, full = TRUE)
}
\arguments{
\item{paceR_db}{The src_mysql connection to the paceR Database.}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}
}
\description{
Get table with group-census data
}
\examples{
getv_CensusMonthly(paceR_db)
}
