% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_paceR_views.R
\name{getv_Individual}
\alias{getv_Individual}
\title{Get table with Individual data}
\usage{
getv_Individual(paceR_db, full = TRUE)
}
\arguments{
\item{paceR_db}{The src_mysql connection to the paceR Database.}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}
}
\description{
Get table with Individual data
}
\examples{
getv_Individual(paceR_db)
}

