% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_paceR_views.R
\name{getv_Phenology}
\alias{getv_Phenology}
\title{Get table with Phenology data}
\usage{
getv_Phenology(paceR_db, full = TRUE)
}
\arguments{
\item{paceR_db}{The src_mysql connection to the paceR Database.}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}
}
\description{
Get table with Phenology data
}
\examples{
getv_Phenology(paceR_db)
}

