% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paceR_call.R
\name{paceR.call}
\alias{paceR.call}
\title{A function to convert allelic depths to genotypes}
\usage{
paceR.call(data = NA, minVafHet = 0.1, maxVafHet = 0.9,
  minDepth = 10, form = "num")
}
\arguments{
\item{data}{Object containing comma seperated allelic depths. Assumes ref,alt format.}

\item{minVafHet}{The minimum variant allele frequency allowed for a variant to be called heterozygous.
A variant with an allele frequency below this will be called homozygous reference.}

\item{maxVafHet}{The maximum variant allele frequency allowed for a variant to be called heterozygous. 
A variant with an allele frequency above this will be called homozygous alternative.}

\item{minDepth}{The minimum number of reads covering a variant to allow it to be called. 
Uncallable regions will be recorded as NA}

\item{form}{Set to either
itemize{
item 'num' -- numeric format where homozygous reference is 0, heterozygous is 1, and homozygous alternative is 2.
item 'char' -- character format where homozygous reference is 0/0, heterozygous is 0/1, and homozygous alternative is 1/1.}}
}
\description{
Pileup Allele CallER: allows you to convert a vector of allelic depths into variant called genotypes
}
\examples{
allelicdepths <- c("5,2", "0", "10,10", "2,28", "15,7")

paceR.call(data=allelicdepths,
 minVafHet=0.1,
 maxVafHet=0.9,
 minDepth=10,
 form='num')
 
[1] NA NA  1  2  1
}
