% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{pheno_avail_indices_sr}
\alias{pheno_avail_indices_sr}
\title{Calculate  availability indices for Santa Rosa using different methods.}
\usage{
pheno_avail_indices_sr(pheno = NULL, smooth = "none", ...)
}
\arguments{
\item{pheno}{The data frame of phenology data from Santa Rosa.}

\item{smooth}{Either "none", "loess", or "gam. The default is "none".}
}
\description{
Calculate  availability indices for Santa Rosa using different methods.
}
\examples{
indices_lo <- pheno_avail_indices_sr(pheno, smooth = "loess")
}
