% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{pheno_prep_fruit_sr}
\alias{pheno_prep_fruit_sr}
\title{Convert a Santa Rosa phenology table from long to wide format and prepare for fruit biomass calculation.}
\usage{
pheno_prep_fruit_sr(pheno = NULL, exclude_species = "", ...)
}
\arguments{
\item{pheno}{The data frame of phenology data from Santa Rosa.}

\item{exclude_species}{A character vector of species codes to exclude.}
}
\description{
Convert a Santa Rosa phenology table from long to wide format and prepare for fruit biomass calculation.
}
\examples{
pheno <- pheno_prep_fruit_sr(ph, exclude_species = c("SPAV", "FUNK"))
}

