% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{pheno_prep_sr}
\alias{pheno_prep_sr}
\title{Convert a Santa Rosa phenology table from long to wide format and prepare for fruit biomass calculation.}
\usage{
pheno_prep_sr(pheno = NULL, exclude_species = "", item = "Fruit",
  maturity = "Mature", ...)
}
\arguments{
\item{pheno}{The data frame of phenology data from Santa Rosa.}

\item{exclude_species}{A character vector of species codes to exclude.}

\item{item}{The food part to focus on. Default is "Fruit".}

\item{maturity}{Maturity of the food part to focus on. Default is "Mature". Only other valid option is "Immature".}
}
\description{
Convert a Santa Rosa phenology table from long to wide format and prepare for fruit biomass calculation.
}
\examples{
pheno <- pheno_prep_sr(ph, exclude_species = c("SPAV", "FUNK"))
}
