% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{plot_biomass_monthly}
\alias{plot_biomass_monthly}
\title{Heatmap plots of total monthly biomass with species combined}
\usage{
plot_biomass_monthly(df = NULL, fill_col = c("#FFFFFF", brewer.pal(9,
  "YlGnBu")))
}
\arguments{
\item{df}{Dataframe of summarized available biomass data created by biomass_monthly_summary.}

\item{fill_col}{A color palette for the fill gradient.}
}
\description{
Heatmap plots of total monthly biomass with species combined
}
\examples{
plot_biomass(biomass_summary_lo)
}

