% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{plot_biomass_species}
\alias{plot_biomass_species}
\title{Generate a heatmap of availabile fruit biomass for each species.}
\usage{
plot_biomass_species(df = NULL, fill_col = c("#FFFFFF",
  RColorBrewer::brewer.pal(9, "YlGnBu")))
}
\arguments{
\item{df}{The data frame of biomass data created by biomass_avail_sr.}

\item{fill_col}{A color palette for the fill gradient.}
}
\description{
Generate a heatmap of availabile fruit biomass for each species.
}
\examples{
plot_biomass_species(biomass_avail_lo)
}
