% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{plot_fpv_dbh}
\alias{plot_fpv_dbh}
\title{Create boxplots of DBHs for trees in FPV data for each species to find size of fruit-producing trees.}
\usage{
plot_fpv_dbh(fpv)
}
\arguments{
\item{fpv}{The FPV data.}
}
\description{
Create boxplots of DBHs for trees in FPV data for each species to find size of fruit-producing trees.
}
\examples{
plot_fpv_dbh(fpv)
}

