% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{plot_pheno_indices}
\alias{plot_pheno_indices}
\title{Generate a heatmap of fruit availability indices.}
\usage{
plot_pheno_indices(df = NULL, fill_col = c("#FFFFFF",
  RColorBrewer::brewer.pal(9, "YlGnBu")))
}
\arguments{
\item{df}{The data frame of index data to plot.}

\item{fill_col}{A color palette for the fill gradient.}
}
\description{
Generate a heatmap of fruit availability indices.
}
\examples{
plot_pheno_indices(indices_lo)
}

