% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sql.R
\name{run_sql_string}
\alias{run_sql_string}
\title{run_sql_string}
\usage{
run_sql_string(sql_string, con = databricks_connect())
}
\arguments{
\item{sql_string}{a string containing sql code to be executed}

\item{con}{a connection object}
}
\value{
tibble
}
\description{
This function takes a database connection (default is \code{paceR::databricks_connect()})
and a \code{sql_string} containing sql to be executed. It returns the result as a tibble.
}
\examples{
\dontrun{
run_sql_string(sql_string = 'SELECT * FROM curated.all.census;')}
}
