% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{sqrt_sign_trans}
\alias{sqrt_sign_trans}
\title{A signed square-root transform.
This transform can be useful when you want to square-root transform
a plot axis to emphasize variation in the small values, but the data
contain negative values.}
\usage{
sqrt_sign_trans()
}
\value{
A transformation
}
\description{
A signed square-root transform.
This transform can be useful when you want to square-root transform
a plot axis to emphasize variation in the small values, but the data
contain negative values.
}
