% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{theme_internal_ticks}
\alias{theme_internal_ticks}
\title{A ggplot2 theme with internal tick marks.}
\usage{
theme_internal_ticks()
}
\value{
A ggplot2 theme.
}
\description{
A ggplot2 theme with internal tick marks.
}
