% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{theme_internal_ticks_x2}
\alias{theme_internal_ticks_x2}
\title{A doubled version of theme_internal_ticks.}
\usage{
theme_internal_ticks_x2()
}
\value{
A ggplot2 theme.
}
\description{
A doubled version of theme_internal_ticks.
}
