% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{theme_journal}
\alias{theme_journal}
\title{A clean ggplot2 theme with no internal grid lines.}
\usage{
theme_journal()
}
\value{
A ggplot2 theme.
}
\description{
A clean ggplot2 theme with no internal grid lines.
}
