% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{tr_to_polys}
\alias{tr_to_polys}
\title{Transforms start- and end-points of transects into polygons (rectangles) with respective width}
\usage{
tr_to_polys(all_tr_points)
}
\arguments{
\item{all_tr_points}{Dataframe bind_rows(horizontal_transect, vertical_transects)}
}
\description{
Transforms start- and end-points of transects into polygons (rectangles) with respective width
}
\examples{
tr_to_polys <- function(all_tr_points)
}
