% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{transect_subset_sr}
\alias{transect_subset_sr}
\title{Get relevant transect data corresponding to pheno species.
Also exclude individual trees that are too small to produce food based on FPVs.}
\usage{
transect_subset_sr(tr = NULL, pheno = NULL, min_dbh = NULL)
}
\arguments{
\item{tr}{PACE transect data.}

\item{pheno}{Phenology data created by pheno_prep_fruit_sr.}

\item{min_dbh}{FPV minimum DBH data created by fpv_get_min_dbh_sr.}
}
\description{
Get relevant transect data corresponding to pheno species.
Also exclude individual trees that are too small to produce food based on FPVs.
}
\examples{
tr_pheno_fpv <- transect_subset_sr(tr, pheno, min_dbh)
}

