% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generate-pace-lifehistory tables.R
\name{get_biography}
\alias{get_biography}
\title{Get table with biography from birth until death, disappearance or end of observation of individuals.}
\usage{
get_biography(pace_db, full = TRUE, projectID = 1)
}
\arguments{
\item{pace_db}{The src_mysql connection to the PACE Database.}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}

\item{projectID}{Option to get data only from specific project (1-7?). Default is 1 (Santa Rosa)}
}
\description{
Get table with biography from birth until death, disappearance or end of observation of individuals.
}
\examples{
get_biography(pace_db)
}

