% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-pace-focaldata-mackenzie.R
\name{get_focaldata_MB}
\alias{get_focaldata_MB}
\title{Get focal behaviour recorded by Mackenzie Bergstrom for her MSc & PhD projects (SubProjectID = 13)}
\usage{
get_focaldata_MB(paceR_db, full = TRUE)
}
\arguments{
\item{paceR_db}{The src_mysql connection to the paceR Database (view-collection).}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}
}
\description{
Get focal behaviour recorded by Mackenzie Bergstrom for her MSc & PhD projects (SubProjectID = 13)
}
\examples{
get_focaldata_MB(paceR_db)
}

