% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{get_horizontal_transects}
\alias{get_horizontal_transects}
\title{Get gps points for horizontal transects}
\usage{
get_horizontal_transects(paceR, tr_full, tr_pt)
}
\arguments{
\item{tr_full}{Dataframe with 'tblVegetationTransect' from Pacelab}

\item{tr_pt}{Dataframe with 'tblVegetationTransectGridPoint' from Pacelab}
}
\description{
Get gps points for horizontal transects
}
\examples{
get_horizontal_transects <- function(paceR, tr_full, tr_pt)
}
