% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-pace-life-history-tables.R
\name{get_infant_table}
\alias{get_infant_table}
\title{Get table with infants including the risk of infanticide as a consequence of
alpha male reversals - NEW FUNCTION!!!}
\usage{
get_infant_table(paceR_db, full = TRUE)
}
\arguments{
\item{paceR_db}{The src_mysql connection to the paceR Database
(view-collection).}

\item{full}{Option to return the full table (TRUE) or just a condensed
version (FALSE). Default is TRUE. CURRENTLY NOT USED.}
}
\description{
This function determines: 1) Alpha male (AM) at the time of birth, 2) days
since AM has been AM, 3) following AM, and 4) days from DOB until next AM
becomes AM
}
\examples{
get_infant_table (paceR_db)
}
