% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load-pace-tables.R
\name{get_monthly_census}
\alias{get_monthly_census}
\title{Get table with group-census data.}
\usage{
get_monthly_census(pace_db, projectID = 1, full = TRUE)
}
\arguments{
\item{pace_db}{The src_mysql connection to the PACE Database.}

\item{projectID}{Option to get data only from specific project (1-7?), standard is set to 1 (Santa Rosa)}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}
}
\description{
Get table with group-census data.
}
\examples{
get_monthly_census(pace_db)
}

