% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_paceR_views.R
\name{getv_AlphaMaleTenure}
\alias{getv_AlphaMaleTenure}
\title{Get table with alpha male tenures.}
\usage{
getv_AlphaMaleTenure(paceR_db, full = TRUE)
}
\arguments{
\item{paceR_db}{The src_mysql connection to the PaceR Database.}

\item{full}{Option to return the full table (TRUE) or just a condensed version (FALSE). Default is TRUE.}
}
\description{
Get table with alpha male tenures.
}
\examples{
getv_AlphaMaleTenure(paceR_db)
}

