% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pheno-functions.R
\name{pheno_fruit_indices_sr}
\alias{pheno_fruit_indices_sr}
\title{Calculate fruit availability indices using different methods.}
\usage{
pheno_fruit_indices_sr(pheno = NULL, method = "raw", ...)
}
\arguments{
\item{pheno}{The data frame of phenology data from Santa Rosa.}

\item{method}{Either "raw", "loess", or "gam. The default is "raw".}
}
\description{
Calculate fruit availability indices using different methods.
}
\examples{
indices_lo <- pheno_fruit_indices_sr(pheno, method = "loess")
}

